project updated by xrx.

+++ src/main/java/com/team01/domain/ProductInfo.java
@@ -77,7 +77,7 @@ public class ProductInfo {
         this.productType = productType;
     }
 
-    public double getProductPrice() {
+    public int getProductPrice() {
         return productPrice;
     }
